requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
cairocffi==1.7.1
CairoSVG==2.8.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
cryptography==45.0.6
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2
django-clearcache==1.2.1
django-crispy-forms==2.4
django-crontab==0.7.1
django-environ==0.12.0
django-filter==25.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
docopt==0.6.2
et_xmlfile==2.0.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kombu==5.5.4
lxml==6.0.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
pipreqs==0.4.13
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==5.2.1
reportlab==4.4.3
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
yarg==0.1.10
